refactor(IoC): :sparkles: introduce inversion of control and dependency injection to plugin
