+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+  "conventionalCommits.scopes": [
+    "IoC"
+  ]
+}
