+++ src/classes/SmartConnectionsPlugin/__template_injector.js
@@ -1,0 +1,9 @@
+// set module export to be a function to support dependency injection
+module.exports = function injectMethod( modifyMe ){
+
+  // inject this method: 
+  modifyMe.prototype.
+    
+  // inject function returns nothing, it just modifies the class
+
+}
