+++ src/classes/SmartConnectionsPlugin/addIcon.js
@@ -1,0 +1,18 @@
+// set module export to be a function to support dependency injection
+// this injector also utilizes the Obsidian class
+module.exports = function injectMethod( Obsidian, modifyMe ){
+
+  // set the addIcon method
+  modifyMe.prototype.addIcon = function(){
+    Obsidian.addIcon("smart-connections", `<path d="M50,20 L80,40 L80,60 L50,100" stroke="currentColor" stroke-width="4" fill="none"/>
+    <path d="M30,50 L55,70" stroke="currentColor" stroke-width="5" fill="none"/>
+    <circle cx="50" cy="20" r="9" fill="currentColor"/>
+    <circle cx="80" cy="40" r="9" fill="currentColor"/>
+    <circle cx="80" cy="70" r="9" fill="currentColor"/>
+    <circle cx="50" cy="100" r="9" fill="currentColor"/>
+    <circle cx="30" cy="50" r="9" fill="currentColor"/>`);
+  }
+
+  // inject function returns nothing, it just modifies the class
+
+}
