+++ src/classes/SmartConnectionsPlugin/add_to_gitignore.js
@@ -1,0 +1,22 @@
+// set module export to be a function to support dependency injection
+module.exports = function injectMethod( modifyMe ){
+
+  // inject this method: 
+  modifyMe.prototype.add_to_gitignore = async function() {
+    if(!(await this.app.vault.adapter.exists(".gitignore"))) {
+      return; // if .gitignore doesn't exist then don't add .smart-connections to .gitignore
+    }
+    let gitignore_file = await this.app.vault.adapter.read(".gitignore");
+    // if .smart-connections not in .gitignore
+    if (gitignore_file.indexOf(".smart-connections") < 0) {
+      // add .smart-connections to .gitignore
+      let add_to_gitignore = "\n\n# Ignore Smart Connections folder because embeddings file is large and updated frequently";
+      add_to_gitignore += "\n.smart-connections";
+      await this.app.vault.adapter.write(".gitignore", gitignore_file + add_to_gitignore);
+      console.log("added .smart-connections to .gitignore");
+    }
+  }
+    
+  // inject function returns nothing, it just modifies the class
+
+}
