+++ src/classes/SmartConnectionsPlugin/output_render_log.js
@@ -1,0 +1,30 @@
+// set module export to be a function to support dependency injection
+module.exports = function injectMethod( modifyMe ){
+
+  // inject this method: 
+  modifyMe.prototype.output_render_log = function() {
+    // if settings.log_render is true
+    if(this.settings.log_render) {
+      if (this.render_log.new_embeddings === 0) {
+        return;
+      }else{
+        // pretty print this.render_log to console
+        console.log(JSON.stringify(this.render_log, null, 2));
+      }
+    }
+
+    // clear render_log
+    this.render_log = {};
+    this.render_log.deleted_embeddings = 0;
+    this.render_log.exclusions_logs = {};
+    this.render_log.failed_embeddings = [];
+    this.render_log.files = [];
+    this.render_log.new_embeddings = 0;
+    this.render_log.skipped_low_delta = {};
+    this.render_log.token_usage = 0;
+    this.render_log.tokens_saved_by_cache = 0;
+  }
+    
+  // inject function returns nothing, it just modifies the class
+
+}
