+++ src/json/defaults.json
@@ -1,0 +1,17 @@
+{
+    "api_key": "",
+    "file_exclusions": "",
+    "folder_exclusions": "",
+    "header_exclusions": "",
+    "path_only": "",
+    "show_full_path": false,
+    "expanded_view": true,
+    "group_nearest_by_file": false,
+    "log_render": false,
+    "log_render_files": false,
+    "skip_sections": false,
+    "smart_chat_model": "gpt-3.5-turbo",
+    "results_count": 30,
+    "view_open": true,
+    "version": ""
+}
